requirements.txt
django==3.1
djangorestframework==3.12
django-filter
requests
django-environ
django-cors-headers
